requirements.txt
streamlit>=1.28.0
langchain-openai>=0.0.8
openai>=1.12.0
python-dotenv>=1.0.0
langchain-community>=0.0.20
pypdf>=3.0.0
docx2txt>=0.8.0
tiktoken>=0.5.0